+++ geo/src/algorithm/contains_properly/mod.rs
@@ -51,6 +51,8 @@ mod line;
 mod line_string;
 mod point;
 mod polygon;
+pub(crate) use polygon::polygon_polygon_inner_loop;
+mod monotone_chain;
 mod rect;
 mod triangle;