+++ geo/src/algorithm/intersects/mod.rs
@@ -68,6 +68,7 @@ mod collections;
 mod coordinate;
 mod line;
 mod line_string;
+mod monotone_chain;
 mod point;
 mod polygon;
 mod rect;
@@ -109,7 +110,7 @@ where
 }
 
 // A cheap bbox check to see if we can skip the more expensive intersection computation
-fn has_disjoint_bboxes<T, A, B>(a: &A, b: &B) -> bool
+pub(crate) fn has_disjoint_bboxes<T, A, B>(a: &A, b: &B) -> bool
 where
     T: CoordNum,
     A: BoundingRect<T>,