+++ geo/src/algorithm/mod.rs
@@ -348,6 +348,12 @@ pub use kmeans::{KMeans, KMeansError, KMeansParams};
 pub mod monotone;
 pub use monotone::{MonoPoly, MonotonicPolygons, monotone_subdivision};
 
+pub mod monotone_chain;
+pub use monotone_chain::{
+    MonotoneChain, MonotoneChainGeometry, MonotoneChainLineString, MonotoneChainMultiLineString,
+    MonotoneChainMultiPolygon, MonotoneChainPolygon, MonotoneChains,
+};
+
 /// Rhumb-line-related algorithms and utils
 pub mod rhumb;
 #[allow(deprecated)]