+++ geo/CHANGES.md
@@ -13,6 +13,11 @@
   - <https://github.com/georust/geo/pull/1487>
 - Fix Euclidean distance fast path for open `LineString`s to consider the last vertex (avoids incorrect `LineString`-to-`LineString` distances for separable geometries).
   - <https://github.com/georust/geo/pull/1499>
+- Add `MonotoneChain`-backed geometry types
+  - Variant of `LineString`, `MultiLineString`, `Polygon`, `MultiPolygon` which are backed by Monotone Chains
+  - Preprocessing cost to construct these types, but provide a significant performance boost for intersects and contains_properly checks
+  - <https://github.com/georust/geo/issues/1466>
+  - <https://github.com/georust/geo/pull/1467>
 
 ## 0.32.0 - 2025-12-05